Fixing the drop rates
Fix provided in order for NPCs to drop more than one item of the same
type.
In the war_world database, creature_loots table the Pct column holds the
percentage for the drop rate of each item. If the Pct was set for
example 200, the item would surely drop but only once.

This is needed as on the original server multiple items of the same type
were indeed dropped. For example when farming, you would put 1 seed,
nourishment and hope to get 1 or 2 seeds + the plant.